requirements.txt
aioredis==1.3.1
asgiref==3.4.1
async-timeout==4.0.2
attrs==21.4.0
autobahn==22.1.1
Automat==20.2.0
backports.entry-points-selectable==1.1.0
certifi==2021.10.8
cffi==1.15.0
channels==3.0.4
channels-redis==3.4.0
cloudinary==1.29.0
constantly==15.1.0
cryptography==36.0.1
daphne==3.0.2
distlib==0.3.3
dj-database-url==0.5.0
Django==3.2.8
django-cors-headers==3.11.0
django-extensions==3.1.5
django-heroku==0.3.1
djangorestframework==3.13.1
exrex==0.10.5
filelock==3.3.0
gunicorn==20.1.0
hiredis==2.0.0
hyperlink==21.0.0
idna==3.3
incremental==21.3.0
msgpack==1.0.4
Pillow==8.3.2
platformdirs==2.4.0
psycopg2==2.9.5
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pyOpenSSL==22.0.0
pytz==2021.3
rstr==3.2.0
service-identity==21.1.0
six==1.16.0
sqlparse==0.4.2
Twisted==22.10.0
txaio==21.2.1
typing_extensions==4.4.0
urllib3==1.26.13
virtualenv==20.8.1
whitenoise==6.0.0
zope.interface==5.4.0
